improved init import and moved base view classes to view module
