+++ visualization/histogram.py
@@ -1,4 +1,4 @@
-from . import MatplotView
+from .view import MatplotView
 from matplotlib import pyplot as plt
 from matplotlib import cm
 from matplotlib.colors import LogNorm
